requirements.txt
Django==4.2.29
djangorestframework==3.16.1
django-redis==5.4.0
django-environ==0.12.0
django-ckeditor-5==0.2.18
uvicorn==0.41.0

channels==4.1.0
channels-redis==4.2.0

whitenoise==6.8.2

psycopg2==2.9.4

pillow==12.0.0

celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.7.0

django-storages==1.14.4
google-cloud-storage==2.18.2